+++ lib/ooxml_parser/pptx_parser/pptx_data/presentation/slide/slide/timing/time_node.rb
@@ -47,7 +47,7 @@ module OoxmlParser
         when 'video'
           timings << :video
         when 'audio'
-          timings << :audio
+          timings << TimeNode.new(:audio, parent: timings).parse(time_node)
         end
       end
       timings